state change

+++ tupan-server/src/main/java/org/entrementes/tupan/repositories/DeviceRegistrationRepository.java
@@ -1,8 +1,10 @@
 package org.entrementes.tupan.repositories;
 
+import java.math.BigInteger;
+
 import org.entrementes.tupan.entity.DeviceRegistrationEntity;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface DeviceRegistrationRepository extends MongoRepository<DeviceRegistrationEntity, Long> {
+public interface DeviceRegistrationRepository extends MongoRepository<DeviceRegistrationEntity, BigInteger> {
 
 }
